+++ Tombolo/server/middlewares/instanceMiddleware.js
@@ -36,23 +36,8 @@ const validateInstanceUpdate = [
   },
 ];
 
-const validateAccessRequest = [
-  body("id").isUUID(4).withMessage("Invalid user id"),
-  body("comment").notEmpty().withMessage("Comment is required"),
-  body("roles").isArray().withMessage("Roles must be an array"),
-  body("applications").isArray().withMessage("Applications must be an array"),
-  (req, res, next) => {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      return res.status(400).json({ errors: errors.array() });
-    }
-    next();
-  },
-];
-
 module.exports = {
   validateInstanceSetting,
   validateInstanceId,
   validateInstanceUpdate,
-  validateAccessRequest,
 };
